wayland: Quiet an ubsan complaint
unsigned char is promoted to int, which lacks the 32nd bit to
make 0xff << 24 work. Explicitly cast to unsigned int to make
it clear what we want to happen.

+++ gdk/wayland/cursor/xcursor.c
@@ -266,10 +266,10 @@ _XcursorReadUInt (XcursorFile *file, XcursorUInt *u)
 
     if ((*file->read) (file, bytes, 4) != 4)
 	return XcursorFalse;
-    *u = ((bytes[0] << 0) |
-	  (bytes[1] << 8) |
-	  (bytes[2] << 16) |
-	  (bytes[3] << 24));
+    *u = ((((unsigned int)(bytes[0])) << 0) |
+	  (((unsigned int)(bytes[1])) << 8) |
+	  (((unsigned int)(bytes[2])) << 16) |
+	  (((unsigned int)(bytes[3])) << 24));
     return XcursorTrue;
 }
